feat: enhance how app loads in default config
- import loaded config from app rather then calling the load over and over

+++ python/templates.py
@@ -3,7 +3,8 @@ import os
 import yaml
 from utils.config import load_config, project_root
 from collections import OrderedDict
-
+from utils.colorized_util import console
+from app import config
 
 def update_frontmatter(frontmatter, key, value):
     frontmatter_dict = frontmatter
@@ -13,7 +14,7 @@ def update_frontmatter(frontmatter, key, value):
 
 def create_new_board(board_config=None):
     # load app configs
-    lowpm_config = load_config()
+    lowpm_config = config
     project_dir = lowpm_config["project_dir"]
 
     # Check if load_config() returned a dictionary
@@ -48,7 +49,7 @@ def create_new_board(board_config=None):
 
     brd_config = {
         "template": board_config["template"] or template_path,
-        "name": f"{board_config["name"]}.{el_type}.md" or f"new.{el_type}.md",
+        "name": f"new.{el_type}.md" if not board_config["name"] else f"{board_config["name"]}.{el_type}.md",
         "path": board_config["path"] or project_dir,
         "frontmatter": frontmatter,
     }
@@ -72,13 +73,17 @@ def create_new_board(board_config=None):
 
     # Create the new file
     new_file_path = os.path.join(brd_config["path"], brd_config["name"])
-    with open(new_file_path, "w") as file:
+
+    if os.path.exists(new_file_path):
+      console.print(f'[bold  dark_orange3 ]WARNING:[/bold dark_orange3] The file [bold]{new_file_path}[/bold] already exists')
+    else:
+      with open(new_file_path, 'w') as file:
         file.write(merged_content)
 
 
 def create_new_page(page_config=None):
     # load app configs
-    lowpm_config = load_config()
+    lowpm_config = config
 
     # Check if load_config() returned a dictionary
     if not isinstance(lowpm_config, dict):
@@ -90,13 +95,15 @@ def create_new_page(page_config=None):
         if lowpm_config["frontmatter"]["page"]
         else None
     )
+    # add doc type to frontmatter
     frontmatter_with_type = update_frontmatter(base_frontmatter, "type", "page")
+    
+    
+
 
     if page_frontmatter:
         frontmatter_with_type.update(page_frontmatter)
-    frontmatter = frontmatter_with_type
-
-    print(frontmatter)
+    frontmatter = frontmatter_with_type    
 
     lowpm_dir = lowpm_config["dir"]
     project_dir = lowpm_config["project_dir"]
@@ -106,13 +113,9 @@ def create_new_page(page_config=None):
         project_root, lowpm_dir, "templates", "template.page.md"
     )
 
-  
-
-    el_type = page_config["type"]
-
     pg_config = {
         "template": page_config["template"] or template_path,
-        "name": f"{page_config["name"]}.md" or f"new.md",
+        "name": f"new.md" if not page_config["name"] else f"{page_config["name"]}.md" ,
         "path": page_config["path"] or project_dir,
         "frontmatter": frontmatter
     }
@@ -137,13 +140,18 @@ def create_new_page(page_config=None):
     # Create the new file
     new_file_path = os.path.join(pg_config["path"], pg_config["name"])
 
-    with open(new_file_path, "w") as file:
+    if os.path.exists(new_file_path):
+      console.print(f'[bold  dark_orange3 ]WARNING:[/bold dark_orange3] The file [bold]{new_file_path}[/bold] already exists')
+    else:
+      with open(new_file_path, 'w') as file:
         file.write(merged_content)
+
+    
 
 
 def create_new_list(list_config=None):
     # load app configs
-    lowpm_config = load_config()
+    lowpm_config = config
     project_dir = lowpm_config["project_dir"]
 
     # Check if load_config() returned a dictionary
@@ -177,7 +185,7 @@ def create_new_list(list_config=None):
 
     li_config = {
         "template": list_config["template"] or template_path,
-        "name": f"{list_config["name"]}.{el_type}.md" or f"new.{el_type}.md",
+        "name": f"new.{el_type}.md" if not list_config["name"] else f"{list_config["name"]}.{el_type}.md",
         "path": list_config["path"] or project_dir,
         "frontmatter": frontmatter,
     }
@@ -201,7 +209,11 @@ def create_new_list(list_config=None):
 
     # Create the new file
     new_file_path = os.path.join(li_config["path"], li_config["name"])
-    with open(new_file_path, "w") as file:
+    
+    if os.path.exists(new_file_path):
+      console.print(f'[bold  dark_orange3 ]WARNING:[/bold dark_orange3] The file [bold]{new_file_path}[/bold] already exists')
+    else:
+      with open(new_file_path, 'w') as file:
         file.write(merged_content)
